please gcc 2.8

+++ sys/arch/hp300/dev/ite.c
@@ -1,4 +1,4 @@
-/*	$OpenBSD: ite.c,v 1.10 1997/04/16 11:56:10 downsj Exp $	*/
+/*	$OpenBSD: ite.c,v 1.11 1998/03/06 17:24:12 millert Exp $	*/
 /*	$NetBSD: ite.c,v 1.38 1997/03/31 07:37:25 scottr Exp $	*/
 
 /*
@@ -521,25 +521,25 @@ itefilter(stat, c)
 	default:
 	case KBD_KEY:
 	        if (!capsmode) {
-			code = kbd_keymap[c];
+			code = kbd_keymap[(int)c];
 			break;
 		}
 		/* FALLTHROUGH */
 
 	case KBD_SHIFT:
-		code = kbd_shiftmap[c];
+		code = kbd_shiftmap[(int)c];
 		break;
 
 	case KBD_CTRL:
-		code = kbd_ctrlmap[c];
+		code = kbd_ctrlmap[(int)c];
 		break;
 		
 	case KBD_CTRLSHIFT:	
-		code = kbd_ctrlshiftmap[c];
+		code = kbd_ctrlshiftmap[(int)c];
 		break;
         }
 
-	if (code == '\0' && (str = kbd_stringmap[c]) != '\0') {
+	if (code == '\0' && (str = kbd_stringmap[(int)c]) != '\0') {
 		while (*str)
 			(*linesw[kbd_tty->t_line].l_rint)(*str++, kbd_tty);
 	} else {
